If you are not satisfiying on this, ask Ubuntu and Dell Support.
Conclusion :
We can conclude that majority of the boot-up-related issues are because of broken driver or system files. We hope that you were quickly able to fix the problem with the help of this article. In case none of the above methods worked, you might need to consider reinstalling Ubuntu on your PC.